benchmark/requirements.txt
openai>=1.0
anthropic>=0.30
google-genai>=1.0
httpx>=0.25
pyyaml>=6.0
pandas>=2.0
matplotlib>=3.8
seaborn>=0.13
tqdm>=4.65
python-dotenv>=1.0
